+++ src/components/UI/buttons/WhiteButtonComponent.vue
@@ -13,7 +13,7 @@
 .white-btn {
   background-color: white; /* 내부 배경색 */
   color: black; /* 글씨 색깔 */
-  border: 2px solid black; /* 검은색 테두리 */
+  border: 2px solid #dedee2; /* 검은색 테두리 */
   padding: 10px 20px;
   font-size: 16px;
   text-align: center;
